fix/#160 : 이용약관으로 이동하는 것을 sideEffect로 처리
