+++ feature/mypage/src/main/java/com/boostcamp/mapisode/mypage/screen/ProfileEditScreen.kt
@@ -65,8 +65,6 @@ fun ProfileEditRoute(
 						Toast.LENGTH_SHORT,
 					).show()
 				}
-
-				is ProfileEditSideEffect.OpenPrivacyPolicy -> TODO()
 			}
 		}
 	}
